Split IBMA into IBMA and ESMA.
Also, move IBMA from its own folder to just a file.
